Updated commands to be work with files and folders

+++ mzalendo/odekro/management/commands/rename_hansard.py
@@ -17,24 +17,42 @@ class Command(BaseCommand):
     help = 'Rename hansard text files based on header information'
     args = '<file>'
 
+    option_list = BaseCommand.option_list + (
+        make_option(
+            '--verbose',
+            action='store_true',
+            dest='verbose',
+            default=False,
+            help='What it says on the tin ;)'),
+        )
+
     def handle(self, *args, **options):
         if not len(args):
             raise CommandError
         
         src = args[0]
-        self.rename(src)
+
+        if not os.path.exists(src)
+            sys.exit(1)
+
+        if os.path.isfile(src):
+            self.rename(src, verbose=options.get('verbose', False))
+        elif os.path.isdir(src):
+            for f in os.listdir(src):
+                self.handle(os.path.join(src, f), **options)
 
     def rename(self, src, name=None, ext=None, verbose=False):
-        if not name:
-            name = self.filename(src)
-        if not ext:
-            ext = src[-4:]
-        
-        dst = os.path.join(os.path.dirname(src), '%s%s' % (name, ext))
-        if verbose:
-            print dst
-        if src != dst:
-            os.rename(src, dst)
+        if os.path.isfile(src):
+            if not name:
+                name = self.filename(src)
+            if not ext:
+                ext = src[-4:]
+            
+            dst = os.path.join(os.path.dirname(src), '%s%s' % (name, ext))
+            if verbose:
+                print dst
+            if src != dst:
+                os.rename(src, dst)
 
     def filename(self, src):
         content = open(src, 'rU').read()
